fix(temporal): two C++ compile errors blocking the node-smol build
node-smol's ninja build failed compiling the temporal C++ port:

1. calendar.h:306 / calendar.cc:391 — 'must use struct tag to refer to
   type Era'. The virtual `Era(...)` method (line 257) shadows the
   `struct Era` type within the class scope, so a bare `const Era&`
   parameter on EraYearToIsoYear resolved to the method, not the type.
   Disambiguate with `const struct Era&` (keeps the method named `Era`,
   matching the temporal_rs `era()` it ports — no upstream divergence).

2. PlainMonthDay.hpp:459 — 'invalid operands (const uint8_t and
   CalendarKind)'. ParsedDate::ToInfra().calendar_kind is a raw uint8_t
   POD field (0 = ISO); CalendarKind is `enum class … : uint8_t` with no
   implicit conversion. Cast the enum to its underlying uint8_t for the
   comparison. Variant analysis: this is the only `.calendar_kind` (raw
   field) vs enum compare in the tree — every other `kind ==` site
   compares an AnyCalendarKind::ToInfra() enum, which is valid.

node-smol copies these from packages/temporal-infra via
prepare-external-sources, so the canonical edits flow into the build.

+++ packages/temporal-infra/include/temporal_rs/PlainMonthDay.hpp
@@ -451,9 +451,14 @@ class PlainMonthDay {
   //                      non-ISO calendars.
   static diplomat::result<std::unique_ptr<PlainMonthDay>, TemporalError>
   from_parsed(const ParsedDate& parsed) {
+    // calendar_kind is a raw uint8_t in the parsed-intermediates POD
+    // (0 = ISO); CalendarKind is `enum class … : uint8_t`, which has no
+    // implicit conversion, so cast the enum to its underlying value to
+    // compare.
     const bool is_iso =
         parsed.ToInfra().calendar_kind ==
-        ::node::socketsecurity::temporal::CalendarKind::kIso;
+        static_cast<uint8_t>(
+            ::node::socketsecurity::temporal::CalendarKind::kIso);
     std::optional<int32_t> reference_year =
         is_iso ? std::optional<int32_t>{}
                : std::optional<int32_t>(parsed.year());
